+++ extensions/openrouter-quick-actions/src/hooks/useCumulativeCost.ts
@@ -1,0 +1,13 @@
+import { useEffect, useState } from "react";
+
+export function useCumulativeCost(cost: number) {
+  const [cumulativeCost, setCumulativeCost] = useState(0);
+
+  useEffect(() => {
+    if (cost > 0) {
+      setCumulativeCost((prev) => prev + cost);
+    }
+  }, [cost]);
+
+  return cumulativeCost;
+}
