+++ extensions/openrouter-quick-actions/src/type.ts
@@ -1,0 +1,14 @@
+export interface RequestMetadata {
+  provider: string;
+  total_cost: number;
+  model: string;
+  latencyMs: number;
+  durationS: string;
+}
+
+export interface ResultViewConfig {
+  prompt: string;
+  model_override: string;
+  provider_sort?: string;
+  toast_title: string;
+}
