+++ extensions/openrouter-quick-actions/src/utils.ts
